more changes

+++ backend/db/migrations/13-create-reviews.js
@@ -34,7 +34,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       stars: {
-        type: Sequelize.INTEGER
+        type: Sequelize.NUMERIC(2, 2)
       },
       createdAt: {
         allowNull: false,
